pyproject.toml
[tool.poetry]
name = "personal-email-agent"
version = "0.1.0"
description = "Gmail classification and organization agent using a local LLM"
authors = ["Scott"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
openai = ">=1.0"
google-api-python-client = ">=2.100"
google-auth-oauthlib = ">=1.0"
google-auth-httplib2 = ">=0.2"
psycopg2-binary = ">=2.9"
pyyaml = ">=6.0"
pydantic = ">=2.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-cov = ">=4.0"
black = ">=23.0"
mypy = ">=1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
